requirements.txt
vit_prisma
open_clip_torch
ipykernel
wandb
sae-lens==2.1.3